Project_Code/requirements.txt
numpy~=1.21.2
scipy~=1.7.1
scikit-learn~=1.0.1
torchvision~=0.11.1
cvxopt~=1.3.0